Changed "Claimed Sections" to "Currently Claimed Sections"
The original wording of "Claimed Sections" doesn't reflect the notion that sections can be claimed and unclaimed dynamically. What matters is who is CURRENTLY claiming any particular sections. When sections are checked back in, then the person "currently claiming" is simply removed (or replaced with new "claims" on other sections).