Finished traces and added an animated comparison.
The remaining sections (39-41) were traced and the project was imported
into Blender using Neuropil Tools. Gamer was used to smooth the
resulting mesh, and it was reoriented to match the original 3d version.

An animation was added to compare the reconstruction with the original.
This commit didn't update the markup yet ... that will be done through
the GitHub interface.